upd Section 4.2, 4.3
Section 4.2 of iGOV  the scopes **openid**, **offline_access** are used in SPID OpenID Connect proposal and they do not consider the other recommended scopes in the iGov profile, namely: **doc**.

Section 4.3 of iGov defines the policy regarding the userinfo object of claim request parameter. In CIE id, we define the policy both for the userinfo and ID Token object.